password funzionano

+++ backend/deleteinvitato_back.php
@@ -18,18 +18,22 @@
                     </div>
                 </div>
                 <?php
-                require_once 'mysql_connect_back.php';
 
-                $query =
-                    'DELETE FROM invito WHERE id_riunione=\'' .
-                    $_GET['id'] .
-                    '\' AND persona=\'' .
-                    $_GET['persona'] .
-                    '\';';
-                //echo $query;
-                $esito = mysqli_query($dbc, $query);
-                if ($esito == true) {
-                    echo '
+                if (hash('sha256', $_POST['password']) != $_SESSION['user_data']['password']){
+                    echo '<h4 class=" alert alert-danger"><strong>Password sbagliata.</strong> operazione non effettuata. Riprova.</h4><a href="../frontend/le%20mie%20riunioni.php">Indietro</a>';
+                } else {
+                    require_once 'mysql_connect_back.php';
+
+                    $query =
+                        'DELETE FROM invito WHERE id_riunione=\'' .
+                        $_GET['id'] .
+                        '\' AND persona=\'' .
+                        $_GET['persona'] .
+                        '\';';
+                    //echo $query;
+                    $esito = mysqli_query($dbc, $query);
+                    if ($esito == true) {
+                        echo '
                                     <h4 class="alert alert-success">
                                         <i class="fas fa-check-circle"></i><strong>  Fatto!</strong> modifica eseguita con successo.
                                     </h4>
@@ -37,14 +41,15 @@
                                         <a class="col-3" href="../frontend/le%20mie%20riunioni.php">torna alle mie riunioni</a>
                                         <a class="col-9" href="../index.php">homepage</a>
                                     </div>';
-                } else {
-                    $ERRORI = mysqli_error($dbc);
-                    echo '
+                    } else {
+                        $ERRORI = mysqli_error($dbc);
+                        echo '
                                     <h4 class="alert alert-danger">
                                         <i class="fas fa-exclamation-circle"></i><strong>  C\'è stato un problema:</br></strong>';
-                    echo $ERRORI, '</h4>';
+                        echo $ERRORI, '</h4>';
+                    }
+                    mysqli_close($dbc);
                 }
-                mysqli_close($dbc);
                 ?>
 
             </div>
